# Add faster color picker drawing