### Description of the feature or fix

This replace the drawing code of the color picker by a faster version.
Basically, instead of computing 360/3 conversion from HSV to RGB with a slow (but precise) function plus 360/3 draw line function call, this one computes 256/3 conversion from HSV to RGB with a fast (but imprecise) function plus 256/3 draw line function call.

On my computer, in the emulator, this decrease the draw time from 30ms to 25ms (with SDL). On emscripten, the FPS goes from 10fps to 15fps.

Please notice that this only affects drawing the color wheel, the angle computations and resulting HSV value is still using the accurate function so this has no impact on the accuracy of the result.

### Checkpoints
- [x] Follow the [styling guide](https://github.com/lvgl/lvgl/blob/master/docs/CODING_STYLE.md)
- [?] Update CHANGELOG.md (don't know if it's required for non-fix PR)
- [x] Update the documentation (not required, there is no change to the API)