@kisvegabor 2021-02-22
@X-Ryl669 Thank you! The 50% performance gain (10->25 FPS) is awesome!

@embeddedt I agree to base it on v8. The color picker is renamed to [lv_colorwheel](https://github.com/lvgl/lvgl/blob/dev/src/extra/widgets/colorwheel/lv_colorwheel.c). The parts affected by this PR hasn't changed much. 

@X-Ryl669 Would you mind sending a new PR for the `dev` brach?