@X-Ryl669 2021-02-22
How can I test it ? Does a simulator exists for dev branch ?