@kisvegabor 2021-02-22
I've just updated the https://github.com/lvgl/lv_sim_eclipse_sdl repo's `dev` branch.